Refactor auth state and account sync to context provider
Moved Firebase auth state change and account synchronization logic from individual pages and the use-account-document hook into the AccountProvider context. This centralizes user/account state management, removes redundant useEffect hooks from multiple pages, and simplifies the use-account-document hook. Also added initial loading state handling to the spinner view.
