+++ src/hooks/use-account-document.ts
@@ -1,7 +1,6 @@
 'use client'
 
 import { doc, getDoc, getFirestore, setDoc } from '@firebase/firestore'
-import { User } from '@firebase/auth'
 import { Account } from '@/libs/interfaces/account.interface'
 import app from '../../firebaseConfig'
 import { useAccount } from '@/components/layout/account-context-provider'
@@ -10,67 +9,6 @@ import { useAccount } from '@/components/layout/account-context-provider'
 export const useAccountDocument = () => {
     const firestore = getFirestore(app)
     const { account, setAccount, user: currentUser } = useAccount()
-    
-    const updateUserAccount = async (user: User | null): Promise<void> => {
-        if (!user) {
-            setAccount(null)
-            return
-        }
-        const accountDocRef = doc(firestore, 'Accounts', user.uid)
-        const docSnap = await getDoc(accountDocRef)
-        
-        let fcmToken
-        if (window.flutter_inappwebview) {
-            fcmToken = await window.flutter_inappwebview.callHandler('requestFcmToken')
-        }
-        if (!docSnap.exists()) {
-            await setDoc(accountDocRef, {
-                uid: user.uid,
-                email: user.email || '',
-                createdAt: new Date(),
-                updatedAt: new Date(),
-                exp: 0,
-                level: 0,
-                fcmToken: fcmToken?.toString() || ''
-            })
-            setAccount(new Account({
-                uid: user.uid,
-                email: user.email || '',
-                createdAt: new Date(),
-                updatedAt: new Date(),
-                exp: 0,
-                level: 0,
-                fcmToken: fcmToken?.toString() || ''
-            }))
-            return
-        }
-        
-        const accountData = docSnap.data()
-        
-        if (fcmToken && fcmToken !== accountData.fcmToken) {
-            await setDoc(accountDocRef, {
-                uid: accountData.uid,
-                email: accountData.email,
-                createdAt: new Date(accountData.createdAt.toDate()),
-                updatedAt: new Date(accountData.updatedAt.toDate()),
-                exp: accountData.exp || 0,
-                level: accountData.level || 0,
-                nickname: accountData.nickname || '',
-                fcmToken: fcmToken
-            })
-        }
-        
-        setAccount(new Account({
-            uid: accountData.uid,
-            email: accountData.email,
-            createdAt: new Date(accountData.createdAt.toDate()),
-            updatedAt: new Date(accountData.updatedAt.toDate()),
-            exp: accountData.exp || 0,
-            level: accountData.level || 0,
-            nickname: accountData.nickname || '',
-            fcmToken: fcmToken || accountData.fcmToken || ''
-        }))
-    }
     
     const updateAccountNickname = async (uid: string, nickname: string): Promise<boolean> => {
         const accountDocRef = doc(firestore, 'Accounts', uid)
@@ -102,5 +40,5 @@ export const useAccountDocument = () => {
     
     const isLoggedIn = !!currentUser && !!account
     
-    return { updateUserAccount, updateAccountNickname, isLoggedIn }
+    return { updateAccountNickname, isLoggedIn }
 }
